requirements.txt
fastapi
uvicorn
sentence-transformers
scikit-learn
numpy
pydantic
python-dotenv